okokok

+++ ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: jpes <jpes@student.42nice.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/28 17:15:26 by jpes              #+#    #+#             */
-/*   Updated: 2023/03/29 15:17:38 by jpes             ###   ########.fr       */
+/*   Updated: 2023/04/25 15:12:43 by jpes             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(const char *s)
+size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i] != '\0')
